// src/app/models/reclamation.model.ts
export class Reclamation {
    recId?: number; // Optional because it's auto-generated by the backend
    description!: string; // Required
    status!: string; // Required
    createdAt?: Date; // Optional
    updatedAt?: Date; // Optional
    user!: { userId: number }; // Use userId instead of id
}